requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.7
pillow==11.1.0
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
